+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+### 0.4.1
+
+  * Renamed `reset` function to `stop`
+  * Fixed `stop` function, now stops animations and clear confettis
+
+### 0.3.9
+
+Release 6 jul 2025
+
+  * Added `triangle` shape
+  * Now `rectangle` shape has a more natural movement
+  * Update dev dependencies
+  * Update README.md
+
+### 0.3.8
+
+Release 3 jul 2025
+
+  * Up version to republish on npm
+
 ### 0.3.7
 
 Release 25 oct 2024